Store cache expiry times per status object
In Zuul v2 there was only a single status object, so storing the cache
expiry time once was easy and sufficient. In zuul v3 we fetch the status
per tenant however we were still only storing the cache expiry time
once.

This meant that the global last update time was being updated when only
one of the status blobs was updated. The pathological case for this
failing would be something like a cache time of 30 seconds and by
fetching the same tenant data every 31 seconds another consumer looking
at a different tenant would never actually see fresh data because the
cache_time was reset to the last fetch.

Fix this so that we store the fetch time in the cache dictionary beside
the data it represents. This makes the existing response creation more
difficult with the existing functions. Rather than fix it in place
unwind some of the function passing logic into something more easy to
rationalize.

Change-Id: I964ccad9f555dab02845b617001ded57fa7f81b5
Signed-off-by: Jamie Lennox <jamielennox@gmail.com>